Refactored FluentValidationTests folder
