+++ Car.Backend/Car.UnitTests/FluentValidationTests/Address/AddressValidatorTest.cs
@@ -1,10 +1,9 @@
-﻿using Car.Data;
-using Car.Data.FluentValidation;
+﻿using Car.Data.FluentValidation;
 using FluentValidation.TestHelper;
 using NUnit.Framework;
 using Xunit;
 
-namespace Car.UnitTests.FluentValidationTests
+namespace Car.UnitTests.FluentValidationTests.Address
 {
     [TestFixture]
     public class AddressValidatorTest
