--- Car.Backend/Car.UnitTests/FluentValidationTests/ChatValidatorTest.cs
@@ -1,58 +1,0 @@
-﻿using Car.Data;
-using Car.Data.Constants;
-using Car.Data.FluentValidation;
-using FluentValidation.TestHelper;
-using NUnit.Framework;
-using Xunit;
-
-namespace Car.UnitTests.FluentValidationTests
-{
-    [TestFixture]
-    public class ChatValidatorTest
-    {
-        private readonly ChatValidator validator;
-
-        public ChatValidatorTest()
-        {
-            validator = new ChatValidator();
-        }
-
-        [Xunit.Theory]
-        [InlineData(0)]
-        [InlineData(-1)]
-        public void Id_IsNotValid_GeneratesValidationError(int value)
-        {
-            validator.ShouldHaveValidationErrorFor(chat => chat.Id, value);
-        }
-
-        [Xunit.Theory]
-        [InlineData(1)]
-        [InlineData(10)]
-        public void Id_IsSpecified_NotGeneratesValidationError(int value)
-        {
-            validator.ShouldNotHaveValidationErrorFor(chat => chat.Id, value);
-        }
-
-        [Xunit.Theory]
-        [InlineData(null)]
-        [InlineData("")]
-        public void Name_IsNull_GeneratesValidationError(string value)
-        {
-            validator.ShouldHaveValidationErrorFor(chat => chat.Name, value);
-        }
-
-        [Fact]
-        public void Name_IsNotValid_GeneratesValidationError()
-        {
-            string longText = new string('*', Constants.StringMaxLength + 1);
-            validator.ShouldHaveValidationErrorFor(chat => chat.Name, longText);
-        }
-
-        [Xunit.Theory]
-        [InlineData("abc")]
-        public void Name_IsSpecified_NotGeneratesValidationError(string value)
-        {
-            validator.ShouldNotHaveValidationErrorFor(chat => chat.Name, value);
-        }
-    }
-}
