+++ Car.Backend/Car.UnitTests/FluentValidationTests/Message/MessageValidatorTest.cs
@@ -1,12 +1,11 @@
 ﻿using System;
-using Car.Data;
 using Car.Data.Constants;
 using Car.Data.FluentValidation;
 using FluentValidation.TestHelper;
 using NUnit.Framework;
 using Xunit;
 
-namespace Car.UnitTests.FluentValidationTests
+namespace Car.UnitTests.FluentValidationTests.Message
 {
     [TestFixture]
     public class MessageValidatorTest
