--- Car.Backend/Car.UnitTests/FluentValidationTests/UpdateUserModelValidatorTest.cs
@@ -1,35 +1,0 @@
-﻿using Car.Data;
-using Car.Domain.FluentValidation;
-using FluentValidation.TestHelper;
-using NUnit.Framework;
-using Xunit;
-
-namespace Car.UnitTests.FluentValidationTests
-{
-    [TestFixture]
-    public class UpdateUserModelValidatorTest
-    {
-        private readonly UpdateUserModelValidator validator;
-
-        public UpdateUserModelValidatorTest()
-        {
-            validator = new UpdateUserModelValidator();
-        }
-
-        [Xunit.Theory]
-        [InlineData(0)]
-        [InlineData(-1)]
-        public void Id_IsNotValid_GeneratesValidationError(int value)
-        {
-            validator.ShouldHaveValidationErrorFor(userModel => userModel.Id, value);
-        }
-
-        [Xunit.Theory]
-        [InlineData(1)]
-        [InlineData(10)]
-        public void Id_IsSpecified_NotGeneratesValidationError(int value)
-        {
-            validator.ShouldNotHaveValidationErrorFor(userModel => userModel.Id, value);
-        }
-    }
-}
