+++ Car.Backend/Car.UnitTests/FluentValidationTests/UserPreferences/UserPreferencesValidatorTest.cs
@@ -1,11 +1,10 @@
-﻿using Car.Data;
-using Car.Data.Constants;
+﻿using Car.Data.Constants;
 using Car.Data.FluentValidation;
 using FluentValidation.TestHelper;
 using NUnit.Framework;
 using Xunit;
 
-namespace Car.UnitTests.FluentValidationTests
+namespace Car.UnitTests.FluentValidationTests.UserPreferences
 {
     [TestFixture]
     public class UserPreferencesValidatorTest
